Fixed game bug and game is in a fully fuctional state
Fixed bug where game automatically fails on second action/click. The
game works as intented at this point and is basically finished, possible
addition to game in future. Added score count to game and need to figure
out how to take that score and store it in a highscore table.